feat: move readEnvVariables to cli
