+++ packages/cli/src/__tests__/utils/read-env-variables.test.ts
@@ -2,7 +2,7 @@ import * as dotenv from 'dotenv';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 
-import { readEnvVariables } from '../../flow-runner/index.js';
+import { readEnvVariables } from '../../utils/read-env-variables.js';
 
 vi.mock('dotenv');
 vi.mock('node:fs');
