+++ packages/cli/src/utils/read-env-variables.ts
